move new hook where other hooks were located. More comment cleanup. Final Round!
